Refactored common scraper code (#193)
* out filenames backwards

* told dev where the output file was saved

* new env for proprietary scrapers that are 'no-browser'

* README.md for two lambdas

* Refactored common scraper code

* No metrics for development

This was polluting the Grafana data when someone successfully ran a scraper locally while the production lambda was failing.

* remove extra comments
