+++ site-scrapers/index.js
@@ -2,7 +2,7 @@ const fs = require("fs");
 
 let scrapers = [];
 
-// args override directory list allowing single site runs, e.g. `node main.js LynnTech`
+// args override directory list allowing single site runs, e.g. `node scraper.js MAImmunizations`
 if (process.argv.length > 2) {
     for (let i = 2; i < process.argv.length; i++) {
         const scraper = require(`./${process.argv[i]}`);
